need to modify a few script view tests then good to go
