+++ public/src/actions/admin_scripts.jsx
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { destroy } from 'redux-form';
-import { SET_SCRIPTS, SET_SCRIPT_CURRENT } from '../reducers/admin_scripts';
+import { SET_SCRIPTS, SET_SCRIPT_CURRENT, SET_SCRIPT_QUESTIONS } from '../reducers/admin_scripts';
 
 export function setScriptsList(scriptList) {
   return {
@@ -13,6 +13,13 @@ export function setCurrentScript(scriptDataObj) {
   return {
     type: SET_SCRIPT_CURRENT,
     payload: scriptDataObj
+  };
+}
+
+export function setScriptQuestions(questions) {
+  return {
+    type: SET_SCRIPT_QUESTIONS,
+    payload: questions
   };
 }
 
@@ -53,6 +60,36 @@ export function postScript(script, questions, history) {
       const customError = {
         message: `error in posting new script to the db: ${err}`,
         name: 'postNewScript function error in script_form.jsx'
+      };
+      throw customError;
+    });
+}
+
+export function fetchScript(id) {
+  return dispatch => axios.get(`/scripts/${id}`)
+    .then((res) => {
+      const { data: scriptData } = res;
+      return dispatch(setCurrentScript(scriptData));
+    })
+    .catch((err) => {
+      const customError = {
+        message: `error fetching script info in script action fetchScript: ${err}`,
+        name: 'script info get request from script component'
+      };
+      throw customError;
+    });
+}
+
+export function fetchScriptQuestions(id) {
+  return dispatch => axios.get(`/scripts/${id}/scriptQuestions/`)
+    .then((res) => {
+      const { data: scriptQuestions } = res;
+      return dispatch(setScriptQuestions(scriptQuestions));
+    })
+    .catch((err) => {
+      const customError = {
+        message: `error fetching script questions in script action fetchScriptQuestions: ${err}`,
+        name: 'script questions get request from script component'
       };
       throw customError;
     });
