+++ public/src/containers/view_script.jsx
@@ -1,13 +1,13 @@
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import { fetchScript, fetchScriptQuestions } from '../actions/script';
+import { fetchScript, fetchScriptQuestions } from '../actions/admin_scripts';
 import ScriptPage from '../components/script/script';
 
 
 function mapStateToProps(state) {
   return {
     current_script: state.admin_scripts.current_script,
-    questions: state.script_questions.questions
+    questions: state.admin_scripts.script_questions
   };
 }
