+++ public/src/reducers/admin_scripts.jsx
@@ -1,9 +1,11 @@
 export const SET_SCRIPTS = 'SET_SCRIPTS';
 export const SET_SCRIPT_CURRENT = 'SET_SCRIPT_CURRENT';
+export const SET_SCRIPT_QUESTIONS = 'SET_SCRIPT_QUESTIONS';
 
 export const defaultScripts = {
   current_script: {},
-  all_scripts: []
+  all_scripts: [],
+  script_questions: []
 };
 
 export function adminScriptsReducer(state = defaultScripts, action) {
@@ -18,6 +20,11 @@ export function adminScriptsReducer(state = defaultScripts, action) {
       return {
         ...state,
         current_script: payload
+      };
+    case SET_SCRIPT_QUESTIONS:
+      return {
+        ...state,
+        script_questions: payload
       };
     default:
       return state;
