+++ public/src/store.jsx
@@ -9,7 +9,6 @@ import { createLogger } from 'redux-logger';
 
 import { authStatusReducer, LOGOUT_USER } from './reducers/login';
 import { accountInfoReducer } from './reducers/account_info';
-import { scriptQuestionsReducer } from './reducers/script';
 import { campaignListReducer } from './reducers/campaign';
 import { adminQuestionsReducer } from './reducers/admin_questions';
 import { adminScriptsReducer } from './reducers/admin_scripts';
@@ -19,7 +18,6 @@ const appReducer = combineReducers({
   form: formReducer,
   auth: authStatusReducer,
   account_info: accountInfoReducer,
-  script_questions: scriptQuestionsReducer,
   admin_campaigns: campaignListReducer,
   admin_questions: adminQuestionsReducer,
   admin_scripts: adminScriptsReducer,
